Move IP Config to Network folder
