+++ Aura Operating System/Aura_OS/Properties/VersionInfo.cs
@@ -2,6 +2,6 @@ namespace Aura_OS
 {
     public class VersionInfo
     {
-        public static string revision = "151220200158";
+        public static string revision = "151220200304";
     }
 }
